Main File
ZENC coin key features are privacy and Onion anonymity based on RAS and AES algorithms. ZENC Coin’s fast transaction confirmation time allows it to move money quickly.
There are 11 layers Onion. Just like that, we have 11 layers of encryption in ZENC coin.
ZENC is a cryptocurrency that allows users to send transparent or anonymous transactions, depending on the requirements. It has developed a unique ZENC pay approach to overcome one of the most important concerns of multi blockchains scaling. In addition, it is the most decentralized blockchain in the crypto market. 
ZENC is a crypto currency that is built on the existing Binance smart chain. It adheres to the BEP-20 token standard. ZENC ensures excellent privacy and security for all transactions utilizing the Zero Coin protocol that allows for lightning-fast transactions with cheaper transaction fees of less than 0.001$ anywhere in the world. It employs the proof-of-stake script method to secure the blockchain network and earns a 30 % APR on coin held. Proof of Stake Mining, payment getaways for internet enterprises, and multiplatform wallets for local merchants are other names. It develops highly sought-after crypto currency platforms on which ZENC Coin be used.
ZENC Coin was released on May 22, 2021, without an ICO pre sale, with 22 blocks mined to secure the premise coin, which is 1,400,000M 0.47% of the whole supply, the last block for proof work at 10,000 blocks, and it is now PURE Proof of Stake.
ZENC coin also open source and decentralized, which means that no one control it.
The project team has developed a unique ZENCStore approach to overcome one of the most important concerns of multi blockchains-scaling. In addition, it is the most decentralized chain on the crypto market.